requirements.txt
yfinance
finnhub-python
sqlite3
python-dotenv
matplotlib
pandas
